+++ changelog.md
@@ -18,6 +18,7 @@
 - support for KT256 (described
   [here](https://datatracker.ietf.org/doc/draft-irtf-cfrg-kangarootwelve/))
 - full set of test vectors for KT256 from the above document
+- benchmarks of marsupial's KT256 bindings
 
 ## [0.0.2] - 2024-06-21